refactor: reestruturando as pastas do projeto
