+++ src/main/java/br/com/solutis/locadora/controller/car/AccessoryController.java
@@ -1,7 +1,7 @@
 package br.com.solutis.locadora.controller.car;
 
-import br.com.solutis.locadora.exception.car.AccessoryException;
-import br.com.solutis.locadora.exception.car.AccessoryNotFoundException;
+import br.com.solutis.locadora.exception.car.accessory.AccessoryException;
+import br.com.solutis.locadora.exception.car.accessory.AccessoryNotFoundException;
 import br.com.solutis.locadora.model.dto.car.AccessoryDto;
 import br.com.solutis.locadora.response.ErrorResponse;
 import br.com.solutis.locadora.service.car.AccessoryService;
