+++ src/main/java/br/com/solutis/locadora/exception/car/accessory/AccessoryException.java
@@ -1,4 +1,4 @@
-package br.com.solutis.locadora.exception.car;
+package br.com.solutis.locadora.exception.car.accessory;
 
 public class AccessoryException extends RuntimeException {
     public AccessoryException(String message, Throwable cause) {
