+++ src/main/java/br/com/solutis/locadora/exception/car/accessory/AccessoryNotFoundException.java
@@ -1,4 +1,4 @@
-package br.com.solutis.locadora.exception.car;
+package br.com.solutis.locadora.exception.car.accessory;
 
 public class AccessoryNotFoundException extends RuntimeException {
     public AccessoryNotFoundException(Long id) {
