+++ src/main/java/br/com/solutis/locadora/exception/car/manufacturer/ManufacturerException.java
@@ -1,4 +1,4 @@
-package br.com.solutis.locadora.exception.car;
+package br.com.solutis.locadora.exception.car.manufacturer;
 
 public class ManufacturerException extends RuntimeException {
     public ManufacturerException(String message, Throwable cause) {
