+++ src/main/java/br/com/solutis/locadora/exception/car/model/ModelException.java
@@ -1,4 +1,4 @@
-package br.com.solutis.locadora.exception.car;
+package br.com.solutis.locadora.exception.car.model;
 
 public class ModelException extends RuntimeException {
     public ModelException(String message, Throwable cause) {
