+++ src/main/java/br/com/solutis/locadora/exception/person/driver/DriverException.java
@@ -1,4 +1,4 @@
-package br.com.solutis.locadora.exception.person;
+package br.com.solutis.locadora.exception.person.driver;
 
 public class DriverException extends RuntimeException {
     public DriverException(String message, Throwable cause) {
