+++ src/main/java/br/com/solutis/locadora/exception/person/employee/EmployeeException.java
@@ -1,4 +1,4 @@
-package br.com.solutis.locadora.exception.person;
+package br.com.solutis.locadora.exception.person.employee;
 
 public class EmployeeException extends RuntimeException {
     public EmployeeException(String message, Throwable cause) {
