+++ src/main/java/br/com/solutis/locadora/exception/person/employee/EmployeeNotFoundException.java
@@ -1,4 +1,4 @@
-package br.com.solutis.locadora.exception.person;
+package br.com.solutis.locadora.exception.person.employee;
 
 public class EmployeeNotFoundException extends RuntimeException {
     public EmployeeNotFoundException(Long id) {
