+++ src/main/java/br/com/solutis/locadora/exception/rent/cart/CartNotFoundException.java
@@ -1,4 +1,4 @@
-package br.com.solutis.locadora.exception.cart;
+package br.com.solutis.locadora.exception.rent.cart;
 
 public class CartNotFoundException extends RuntimeException {
     public CartNotFoundException(Long id) {
