+++ src/main/java/br/com/solutis/locadora/exception/rent/insurace/InsurancePolicyException.java
@@ -1,4 +1,4 @@
-package br.com.solutis.locadora.exception.rent;
+package br.com.solutis.locadora.exception.rent.insurace;
 
 public class InsurancePolicyException extends RuntimeException {
     public InsurancePolicyException(String message, Throwable cause) {
