+++ src/main/java/br/com/solutis/locadora/model/entity/rent/Rent.java
@@ -1,7 +1,6 @@
 package br.com.solutis.locadora.model.entity.rent;
 
 import br.com.solutis.locadora.model.entity.car.Car;
-import br.com.solutis.locadora.model.entity.cart.Cart;
 import br.com.solutis.locadora.model.entity.person.Driver;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.persistence.*;
