+++ src/main/java/br/com/solutis/locadora/repository/rent/CartRepository.java
@@ -1,6 +1,6 @@
-package br.com.solutis.locadora.repository.cart;
+package br.com.solutis.locadora.repository.rent;
 
-import br.com.solutis.locadora.model.entity.cart.Cart;
+import br.com.solutis.locadora.model.entity.rent.Cart;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
