+++ src/main/java/br/com/solutis/locadora/service/car/CarService.java
@@ -2,7 +2,7 @@ package br.com.solutis.locadora.service.car;
 
 import br.com.solutis.locadora.exception.car.CarException;
 import br.com.solutis.locadora.exception.car.CarNotFoundException;
-import br.com.solutis.locadora.exception.car.ManufacturerNotFoundException;
+import br.com.solutis.locadora.exception.car.manufacturer.ManufacturerNotFoundException;
 import br.com.solutis.locadora.mapper.GenericMapper;
 import br.com.solutis.locadora.model.dto.car.CarDto;
 import br.com.solutis.locadora.model.entity.car.Accessory;
